refactor(di): AppEnvironment의 imageRepository를 프로토콜 타입으로 전환
- ImageRepository 프로토콜에 Sendable 제약 추가, imageUpdatedPublisher를 요구사항으로 노출
- ImageUpdateType을 ImageRepositoryImpl 중첩 타입에서 Domain 모듈 최상위 타입으로 이동
- imageRepository 속성 타입을 ImageRepositoryImpl(구체)에서 ImageRepository(프로토콜)로 변경
- 프로토콜은 기본값 인자를 가질 수 없어, createImage 호출부 1곳에 originalImageID·thumbnailID 인자를 명시
